requirements.txt
discord.py==1.7.3
python-dotenv==0.19.1
python-dateutil==2.8.2